requirements.txt
django == 1.10
pillow
django-queryset-csv
django-imagekit
django-appconf
django-tinymce4-lite
django-filebrowser-no-grappelli
django-formtools

